Bump up the version to 0.2.1 and set sdk version header

+++ test/tests_commands.py
@@ -18,9 +18,10 @@ class ParserTest(unittest.TestCase):
     def setUp(self):
         self.parser = argparser.create_argparser()
 
-    def assert_file_created(self, path):
+    def assert_file_created(self, path, keep: bool = False):
         self.assertTrue(os.path.isfile(path))
-        os.remove(path)
+        if not keep:
+            os.remove(path)
 
     def test_no_command(self):
         result = subprocess.run(['furiosa'], capture_output=True)
@@ -31,7 +32,7 @@ class ParserTest(unittest.TestCase):
                                  '-d',
                                  '-v',
                                  'compile', self.mnist_model,
-                                ], capture_output=True)
+                                 ], capture_output=True)
         self.assertEqual(0, result.returncode, result.stderr)
         self.assertIn('output.enf has been generated', str(result.stdout))
 
@@ -40,7 +41,7 @@ class ParserTest(unittest.TestCase):
                                  '-d',
                                  '-v',
                                  'compile', self.mnist_model,
-                                ], capture_output=True)
+                                 ], capture_output=True)
         self.assertEqual(0, result.returncode, result.stderr)
         self.assertIn('output.enf has been generated', str(result.stdout))
 
@@ -119,8 +120,9 @@ class ParserTest(unittest.TestCase):
                                 capture_output=True)
         self.assertEqual(4, result.returncode)
         self.assertIn('ERROR: fail to compile', str(result.stderr))
-        self.assertIn('test_data/MNISTnet_uint8_quant_without_softmax.tflite (http_status: 501, error_code: CompilationFailed, message: Currently, our compiler is not able to compile this model)',
-                     str(result.stderr))
+        self.assertIn(
+            'test_data/MNISTnet_uint8_quant_without_softmax.tflite (http_status: 501, error_code: CompilationFailed, message: Currently, our compiler is not able to compile this model)',
+            str(result.stderr))
 
     def test_perfeye(self):
         output_path = '/tmp/{}.html'.format(uuid.uuid4())
@@ -156,7 +158,7 @@ class ParserTest(unittest.TestCase):
                                  '-v',
                                  'perfeye',
                                  self.mnist_model,
-                                 #'--target-npu-spec', self.target_npu_spec,
+                                 # '--target-npu-spec', self.target_npu_spec,
                                  '--config', self.compiler_config,
                                  '-o', output_path,
                                  ],
@@ -165,30 +167,29 @@ class ParserTest(unittest.TestCase):
         self.assertIn('{} has been generated'.format(output_path), str(result.stdout))
         self.assert_file_created(output_path)
 
-    def test_optimize(self):
-        output_path = '/tmp/{}.onnx'.format(uuid.uuid4())
+    def test_optimize_and_quantize(self):
+        optimized_model_path = '/tmp/{}.onnx'.format(uuid.uuid4())
         result = subprocess.run(['furiosa',
                                  '-v',
                                  'optimize',
                                  self.test_onnx_model,
-                                 '-o', output_path,
+                                 '-o', optimized_model_path,
                                  ],
                                 capture_output=True)
         self.assertEqual(0, result.returncode, result.stderr)
-        self.assert_file_created(output_path)
 
-    def test_quantize(self):
-        output_path = '/tmp/{}.onnx'.format(uuid.uuid4())
+        quantized_model_path = '/tmp/{}.onnx'.format(uuid.uuid4())
         result = subprocess.run(['furiosa',
-                                 '-v',
                                  'quantize',
-                                 self.test_onnx_model,
+                                 optimized_model_path,
+                                 '-o', quantized_model_path,
                                  '--dynamic-ranges', self.test_dynamic_ranges,
-                                 '-o', output_path,
                                  ],
                                 capture_output=True)
         self.assertEqual(0, result.returncode, result.stderr)
-        self.assert_file_created(output_path)
+
+        self.assert_file_created(optimized_model_path)
+        self.assert_file_created(quantized_model_path)
 
     def test_build_calibration_model(self):
         output_path = '/tmp/{}.onnx'.format(uuid.uuid4())
